+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Termshark now supports multiple live captures/interfaces on the command-line e.g. `termshark -i eth0 -i eth1`
 - Termshark's packet hex view displays a scrollbar if the data doesn't fit in the space available.
 - Termshark can show a capture file's properties using the capinfos binary (bundled with tshark).
+- Termshark now supports [extcap interfaces](https://tshark.dev/capture/sources/extcap_interfaces/) by default. 
 
 ## [2.0.3] - 2019-12-23